+++ src/Symfony/Component/HttpKernel/CHANGELOG.md
@@ -11,6 +11,7 @@ CHANGELOG
  * added the possibility to change the query string parameter used by `UriSigner`
  * deprecated `LazyLoadingFragmentHandler::addRendererService()`
  * deprecated `Extension::addClassesToCompile()`
+ * deprecated `Psr6CacheClearer::addPool()`
 
 3.2.0
 -----